% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nm_1_noise_close_500}
\alias{nm_1_noise_close_500}
\title{A Mixture of Two Close Normal Distributions with 1% of Points Randomly Substituted
by High Atypical Points - 500 Observations}
\format{
A matrix with 500 rows and 3 variables. The first two variables make
  the bivariate data, while the last variable refers to cluster memberships.
  The first 150 rows belong to cluster 1, and the last 350 rows belong to cluster
  2
\describe{
  \item{d1}{variable 1.}
  \item{d2}{variable 2.}
  \item{cluster}{cluster memberships}
}
}
\source{
Punzo, A. and McNicholas, P.D., 2016. Parsimonious mixtures of multivariate
  contaminated normal distributions. \emph{Biometrical Journal, 58}(6), pp.1506-1537.
}
\usage{
nm_1_noise_close_500
}
\description{
A simulated mixture of two close normal distributions with 1% of points randomly
substituted by high atypical points. Refer to Punzo and McNicholas (2016) for
more information about the underlying distribution that generates this data set.
}
\keyword{datasets}
