% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nm_5_noise_far_100}
\alias{nm_5_noise_far_100}
\title{A Mixture of Two Far Normal Distributions with 5% of Points Randomly Substituted
by Noise - 100 Observations}
\format{
A matrix with 100 rows and 3 variables. The first two variables make
  the bivariate data, while the last variable refers to cluster memberships.
  The first 30 rows belong to cluster 1, and the last 70 rows belong to cluster
  2
\describe{
  \item{d1}{variable 1.}
  \item{d2}{variable 2.}
  \item{cluster}{cluster memberships}
}
}
\source{
Punzo, A. and McNicholas, P.D., 2016. Parsimonious mixtures of multivariate
  contaminated normal distributions. \emph{Biometrical Journal, 58}(6), pp.1506-1537.
}
\usage{
nm_5_noise_far_100
}
\description{
A simulated mixture of two far normal distributions with 1% of points randomly
substituted by noise. Refer to Punzo and McNicholas (2016) for more information
about the underlying distribution that generates this data set.
}
\keyword{datasets}
