% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tm_far_100}
\alias{tm_far_100}
\title{A Mixture of Two Far Student's \eqn{t} Distributions - 100 Observations}
\format{
A matrix with 500 rows and 4 variables. The first two variables make
  the bivariate data; the last variable refers to cluster memberships; and the
  last variable refers to outlier indication (1 means outlier, 0 otherwise).
  The first 30 rows belong to cluster 1, and the last 70 rows belong to cluster
  2
\describe{
  \item{d1}{variable 1.}
  \item{d2}{variable 2.}
  \item{cluster}{cluster memberships}
  \item{outlier}{outlier indication}
}
}
\source{
Punzo, A. and McNicholas, P.D., 2016. Parsimonious mixtures of multivariate
  contaminated normal distributions. \emph{Biometrical Journal, 58}(6), pp.1506-1537.
}
\usage{
tm_far_100
}
\description{
A simulated mixture of two far Student's \eqn{t} distributions. Refer to Punzo
and McNicholas (2016) for more information about the underlying distribution
that generates this data set.
}
\keyword{datasets}
