% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UScost}
\alias{UScost}
\title{US Cost of Living Indices in 2019 Data Set}
\format{
A data frame with 50 rows and 7 variables. There are no missing values
\describe{
  \item{Abbr}{State abbreviation.}
  \item{State}{State name.}
  \item{Grocery}{Grocery index.}
  \item{Housing}{Housing index.}
  \item{Utilities}{Utilities index}
  \item{Transportation}{Transporation index.}
  \item{Misc}{Miscellaneous index}
}
}
\source{
\url{https://worldpopulationreview.com}
}
\usage{
UScost
}
\description{
The data set contains the 2019 cost of living indices of 50 states in five different categories: grocery,
housing, transportation, utilities, and miscellaneous (Washington DC is not included). The indices
are calculated by first determining the average cost of living in the United States to
be used as a baseline set at 100. States are then measured against this baseline. For example,
a state with a cost of living index of 200 is twice as expensive as the national average.
}
\keyword{datasets}
