% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCNM.R
\name{MCNM}
\alias{MCNM}
\title{Multivariate Contaminated Normal Mixture (MCNM)}
\usage{
MCNM(
  X,
  G,
  max_iter = 20,
  epsilon = 0.01,
  init_method = c("kmedoids", "kmeans", "hierarchical", "manual", "soft", "hard"),
  equal_prop = FALSE,
  identity_cov = FALSE,
  eta_min = 1.001,
  show_progress = TRUE,
  manual_clusters = NULL
)
}
\arguments{
\item{X}{An \eqn{n} by \eqn{d} matrix or data frame where \eqn{n} is the number of
observations and \eqn{d} is the number of columns or variables.}

\item{G}{The number of clusters.}

\item{max_iter}{(optional) A numeric value giving the maximum number of
iterations each EM algorithm is allowed to use; 20 by default.}

\item{epsilon}{(optional) A number specifying the epsilon value for the
Aitken-based stopping criterion used in the EM algorithm: 0.01 by default.}

\item{init_method}{(optional) A string specifying the method to initialize
the EM algorithm. "kmedoids" clustering is used by default. Alternative
methods include "kmeans", "hierarchical", "manual", "soft", "hard". When
"manual" is chosen, a vector \code{manual_clusters} of length \eqn{n} must
be specified.}

\item{equal_prop}{(optional) A logical value indicating whether mixing
proportions should be equal at initialization of the EM algorithm; FALSE by
default.}

\item{identity_cov}{(optional) A logical value indicating whether covariance
matrices should be initialized as identity matrices; FALSE by default.}

\item{eta_min}{(optional) A numeric value close to 1 to the right specifying
the minimum value of eta; 1.001 by default.}

\item{show_progress}{(optional) A logical value indicating whether the
fitting progress should be displayed; TRUE by default.}

\item{manual_clusters}{A vector of length \eqn{n} that specifies the initial
cluster memberships of the user when \code{init_method} is set to "manual".
Both numeric and character vectors are acceptable. This argument is NULL by
default, so that it is ignored whenever other given initialization methods
are chosen.}
}
\value{
An object of class \code{MixtureMissing} with:
  \item{model}{The model used to fit the data set}
  \item{pi}{Mixing proportions.}
  \item{mu}{Component mean vectors.}
  \item{sigma}{Component covariance matrices.}
  \item{alpha}{Component proportions of good observations.}
  \item{eta}{Component degrees of contamination.}
  \item{z_tilde}{An \eqn{n} by \eqn{G} matrix where each row indicates the expected
    probabilities that the corresponding observation belongs to each cluster.}
  \item{v_tilde}{An \eqn{n} by \eqn{G} matrix where each row indicates the expected
    probabilities that the corresponding observation is outlying with respect
    to each cluster.}
  \item{clusters}{A numeric vector of length \eqn{n} indicating cluster
    memberships determined by the model.}
  \item{outliers}{A logical vector of length \eqn{n} indicating observations that are ouliers.}
  \item{data}{The original data set if it is complete; otherwise, this is
    the data set with missing values imputed by appropriate expectations.}
  \item{complete}{A logical vector of length \eqn{n} indicating which observation(s)
    have no missing values.}
  \item{npar}{The breakdown of the number of parameters to estimate.}
  \item{max_iter}{Maximum number of iterations allowed in the EM algorithm.}
  \item{iter_stop}{The actual number of iterations needed when fitting the
    data set.}
  \item{final_lik}{The final value of likelihood.}
  \item{final_loglik}{The final value of log-likelihood.}
  \item{lik}{All the values of likelihood.}
  \item{loglik}{All the values of log-likelihood.}
  \item{AIC}{Akaike information criterion.}
  \item{BIC}{Bayesian information criterion.}
  \item{KIC}{Kullback information criterion.}
  \item{KICc}{Corrected Kullback information criterion.}
  \item{AIC3}{Modified AIC.}
  \item{CAIC}{Bozdogan's consistent AIC.}
  \item{AICc}{Small-sample version of AIC.}
  \item{ent}{Entropy}
  \item{ICL}{Integrated Completed Likelihood criterion.}
  \item{AWE}{Approximate weight of evidence.}
  \item{CLC}{Classification likelihood criterion.}
  \item{init_method}{The initialization method used in model fitting.}
}
\description{
Carries out model-based clustering using a multivariate contaminated normal
mixture (MCNM). The function will determine itself if the data set is
complete or incomplete and fit the appropriate model accordingly. When using
this function, the data set must be at least bivariate, and missing values
are assumed to be missing at random (MAR).
}
\examples{

data('nm_5_noise_close_100')

#++++ With no missing values ++++#

X <- nm_5_noise_close_100[, 1:2]
mod <- MCNM(X, G = 2, init_method = 'kmedoids', max_iter = 10)

summary(mod)
plot(mod)

#++++ With missing values ++++#

set.seed(1234)

X <- hide_values(nm_5_noise_close_100[, 1:2], prop_cases = 0.1)
mod <- MCNM(X, G = 2, init_method = 'kmedoids', max_iter = 10)

summary(mod)
plot(mod)

}
\references{
Punzo, A. and McNicholas, P.D., 2016. Parsimonious mixtures of multivariate
  contaminated normal distributions. \emph{Biometrical Journal, 58}(6), pp.1506-1537. \cr \cr
Tong, H. and Tortora, C. (2020). Cluster analysis and outlier detection with
  missing data. \emph{arXiv preprint arXiv:2012.05394.}
}
